make deleteUser also delete all of the users posts and sensitive data

+++ controllers/usersController.js
@@ -1,6 +1,8 @@
 const mongoose = require("mongoose");
 const { User, SensitiveData } = require("../models/users");
+const Post = require("../models/posts");
 const bcrypt = require("bcryptjs");
+const { default: axios } = require("axios");
 
 // get all users
 const getUsers = async (req, res) => {
@@ -69,11 +71,22 @@ const updateUser = async (req, res) => {
 // delete user by id
 const deleteUser = async (req, res) => {
 	const { id } = req.params;
-	if (!mongoose.Types.ObjectId.isValid(id)) {
-		return res.status(404).send(`No user with id: ${id}`);
+	try {
+		console.log("Deleting user with ID:", id);
+		if (!mongoose.Types.ObjectId.isValid(id)) {
+			return res.status(404).send(`No user with id: ${id}`);
+		}
+		const user = await User.findById(id);
+
+		if (!user) return res.status(404).json({ message: `User with id ${id} not found.` });
+		await Post.deleteMany({ user: user._id });
+		await User.findByIdAndDelete(id);
+		await SensitiveData.findByIdAndDelete(user.sensitiveData);
+
+		res.status(200).json({ message: "User and their posts deleted successfully." });
+	} catch (err) {
+		res.status(500).json({ message: err.message });
 	}
-	await User.findByIdAndDelete(id);
-	res.status(200).json({ message: "User deleted successfully." });
 };
 
 // get followers by userTag
